requirements.txt
fastapi
uvicorn[standard]
browser-use
playwright
langchain-google-genai
google-generativeai
notion-client
python-dotenv
aiohttp
httpx
